Migrate shard processor to its own class
